refactor: establish observer backend boundary
