+++ .click/evidence-shards.json
@@ -149,6 +149,7 @@
               "tests.test_click_hook",
               "tests.test_click_host_coverage",
               "tests.test_click_host_router",
+              "tests.test_click_observer_backend",
               "tests.test_click_shadow_dashboard",
               "tests.test_click_shadow_intelligence",
               "tests.test_runner_state_recovery",
@@ -161,6 +162,7 @@
             "tests/test_click_hook.py",
             "tests/test_click_host_coverage.py",
             "tests/test_click_host_router.py",
+            "tests/test_click_observer_backend.py",
             "tests/test_click_shadow_dashboard.py",
             "tests/test_click_shadow_intelligence.py",
             "tests/test_runner_state_recovery.py"
