+++ GUARD_CLASSIFICATION.md
@@ -47,7 +47,7 @@ not claim that every hard gate has already been moved to its target layer.
 | Legacy verification class-unit values | `hooks/click_verification_policy.py` and `hooks/click_verification_meter.py` | Compatibility data only — they do not produce runtime advice, prove cost or sufficiency, or participate in receipt authority |
 | Requiring a mutation after a fixed number of otherwise legitimate failed retries | argv handling in `hooks/click_verification.py` and Browser retry handling | Advisory only in v0.30.0; fresh authorization remains distinct from replay of an active or consumed runner |
 | Browser repeat/retry guidance, preferred timing thresholds and interaction-history depth | `hooks/click_browser_advisory.py`; bounded history compaction in Browser preparation | Advisory only — normalized repeats and long timed interactions remain receipt-bound and allowed; old per-input guidance is compacted instead of blocking a new call |
-| Linux Shadow Observer collection and status | schema in `hooks/click_dependency_cache.py`, collection in `hooks/click_dependency_trace.py`, and isolated storage in `hooks/click_verification.py` | Telemetry only — every record is non-authoritative, never permits reuse, never changes check execution or result, and is excluded from completion receipts |
+| Cross-platform Shadow Observer selection, Linux collection and status | schema in `hooks/click_dependency_cache.py`; compatibility facade in `hooks/click_dependency_trace.py`; capability selection in `hooks/click_observer_backend.py`; common result and lifecycle helpers in `hooks/click_observer_common.py`; Linux collection in `hooks/click_observer_linux.py`; and isolated storage in `hooks/click_verification.py` | Telemetry only — every record is non-authoritative, never permits reuse, never changes check execution or result, and is excluded from completion receipts; unimplemented platform backends report unavailable rather than simulating observation |
 | Shadow input fingerprints, pre-run predictions and post-run evaluations | `hooks/click_shadow_intelligence.py`; preparation and result wiring in `hooks/click_verification.py` | Telemetry only — every real check still runs exactly once; prediction agreement is measured after the run and never becomes evidence, reuse, approval, completion, or receipt authority |
 | Local Evidence Map and Shadow ROI dashboard | sanitized projection in `hooks/click_shadow_intelligence.py`, read-only loopback server in `hooks/click_shadow_dashboard.py`, and lifecycle routing in `hooks/click_lifecycle.py` | Explanatory view only — it exposes no raw state or state-changing endpoint, and start/stop cannot change revision, evidence, or completion |
 | Plain-language-first Guarded contract presentation and optional original disclosure | `hooks/click_contract.py`, Skills, documentation, and semantic grader | User-facing explanation policy — the easy body must preserve the canonical meaning, while digest, ID, later-turn approval, and mutation authority remain Core |
