+++ hooks/click_dependency_trace.py
@@ -1,683 +1,66 @@
 #!/usr/bin/env python3
-"""Linux-only, non-authoritative dependency tracing for argv verification.
+"""Compatibility facade and cross-platform selector for Shadow Observer v1.
 
-This leaf executes one already-authorized command and returns its exit status
-beside a bounded Shadow Observer v1 record. Shadow data is never converted to
-the runtime dependency observation used by evidence reuse.
+Callers keep the established ``click_dependency_trace`` surface while common
+record handling, backend selection, and Linux collection have independent
+owners. Unsupported platforms execute the real command once without a
+collector and record honest unavailable telemetry.
 """
 
 from __future__ import annotations
 
-import ast
 from collections.abc import Callable, Mapping, Sequence
-from dataclasses import dataclass
-from functools import lru_cache
-import hashlib
-import os
 from pathlib import Path
 import platform
-import re
-import select
 import subprocess
-import tempfile
-import threading
-import time
 from typing import Any
 
 if __package__:
-    from . import click_dependency_cache, click_process
+    from . import (
+        click_observer_backend,
+        click_observer_common,
+        click_observer_linux,
+    )
 else:  # Executed directly from the bundled hooks directory.
-    import click_dependency_cache
-    import click_process
+    import click_observer_backend
+    import click_observer_common
+    import click_observer_linux
 
 
-SHADOW_STATE_VERSION = 1
-SHADOW_STATE_FIELD = "shadow_observer"
-MAX_SHADOW_STATE_RECORDS = 256
-MAX_RAW_TRACE_BYTES = 4 * 1024 * 1024
-MAX_BACKEND_VERSION_BYTES = 4_096
-STRACE_STRING_LIMIT = click_dependency_cache.MAX_SHADOW_OBSERVER_PATH_BYTES
-STRACE_TRACE_EXPRESSION = "trace=%file,%process,getdents64"
+SHADOW_STATE_VERSION = click_observer_common.SHADOW_STATE_VERSION
+SHADOW_STATE_FIELD = click_observer_common.SHADOW_STATE_FIELD
+MAX_SHADOW_STATE_RECORDS = click_observer_common.MAX_SHADOW_STATE_RECORDS
+MAX_RAW_TRACE_BYTES = click_observer_linux.MAX_RAW_TRACE_BYTES
+MAX_BACKEND_VERSION_BYTES = click_observer_linux.MAX_BACKEND_VERSION_BYTES
+STRACE_STRING_LIMIT = click_observer_linux.STRACE_STRING_LIMIT
+STRACE_TRACE_EXPRESSION = click_observer_linux.STRACE_TRACE_EXPRESSION
 
-_PID_BRACKET = re.compile(r"^\[pid\s+(\d+)\]\s+")
-_PID_PLAIN = re.compile(r"^(\d+)\s+")
-_CALL = re.compile(r"^([A-Za-z0-9_]+)\((.*)\)\s+=\s+(.+)$")
-_QUOTED = re.compile(r'"(?:\\.|[^"\\])*"')
-_FD_PATH = re.compile(r"(?:-?\d+|AT_FDCWD)<([^<>]+)>")
-_VERSION = re.compile(r"\bstrace\s+--\s+version\s+([A-Za-z0-9][A-Za-z0-9._+-]{0,63})")
-_RETURN_INTEGER = re.compile(r"^(-?\d+)")
-_DIGEST = re.compile(r"^[0-9a-f]{64}$")
+BackendCapability = click_observer_backend.BackendCapability
+ObserverBackend = click_observer_backend.ObserverBackend
+ShadowExecution = click_observer_common.ShadowExecution
+ParsedTrace = click_observer_linux.ParsedTrace
 
-_EXEC_CALLS = frozenset({"execve", "execveat"})
-_OPEN_CALLS = frozenset({"open", "openat", "openat2"})
-_METADATA_CALLS = frozenset(
-    {
-        "access",
-        "faccessat",
-        "faccessat2",
-        "lstat",
-        "newfstatat",
-        "readlink",
-        "readlinkat",
-        "stat",
-        "statx",
-    }
-)
-_DIRECTORY_CALLS = frozenset({"getdents64"})
-_CHDIR_CALLS = frozenset({"chdir", "fchdir"})
-_CHILD_CALLS = frozenset({"clone", "clone3", "fork", "vfork"})
-_IGNORED_PROCESS_CALLS = frozenset(
-    {
-        "exit",
-        "exit_group",
-        "getpgrp",
-        "getpgid",
-        "getsid",
-        "kill",
-        "pidfd_open",
-        "pidfd_send_signal",
-        "prctl",
-        "setpgid",
-        "setsid",
-        "tgkill",
-        "tkill",
-        "wait4",
-        "waitid",
-        "waitpid",
-    }
-)
+FallbackExecutor = click_observer_common.FallbackExecutor
+BackendResolver = click_observer_linux.BackendResolver
+FileDigester = click_observer_linux.FileDigester
+BackendProbe = click_observer_linux.BackendProbe
+SpawnArgv = click_observer_linux.SpawnArgv
 
+select_backend = click_observer_backend.select_backend
+run_unobserved = click_observer_common.run_unobserved
+combine_records = click_observer_common.combine_records
+fresh_state = click_observer_common.fresh_state
+state_is_valid = click_observer_common.state_is_valid
+records_from_verification = click_observer_common.records_from_verification
+store_records = click_observer_common.store_records
+advisory = click_observer_common.advisory
 
-FallbackExecutor = Callable[[], int]
-BackendResolver = Callable[..., tuple[str | None, str]]
-FileDigester = Callable[[Path], str]
-BackendProbe = Callable[[str], str]
-SpawnArgv = Callable[..., subprocess.Popen[Any]]
-
-
-@dataclass(frozen=True, slots=True)
-class ParsedTrace:
-    inputs: tuple[dict[str, Any], ...]
-    external_input_count: int
-    unresolved_event_count: int
-    child_process_count: int
-    process_tree_complete: bool
-    root_exec_observed: bool
-
-
-@dataclass(frozen=True, slots=True)
-class ShadowExecution:
-    exit_code: int
-    record: dict[str, Any]
-
-
-@dataclass(slots=True)
-class _PipeCapture:
-    data: bytes = b""
-    truncated: bool = False
-    failed: bool = False
-
-
-def _bounded_add(left: int, right: int) -> int:
-    return min(click_dependency_cache.MAX_JSON_SAFE_INTEGER, left + right)
-
-
-def _file_digest(path: Path) -> str:
-    hasher = hashlib.sha256()
-    try:
-        with path.open("rb") as handle:
-            while True:
-                chunk = handle.read(1024 * 1024)
-                if not chunk:
-                    break
-                hasher.update(chunk)
-    except OSError:
-        return ""
-    return hasher.hexdigest()
-
-
-@lru_cache(maxsize=8)
-def probe_strace_version(executable: str) -> str:
-    """Return a bounded strace version, or an empty value on probe failure."""
-    try:
-        result = click_process.run_argv(
-            [executable, "--version"],
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-            timeout=2.0,
-        )
-    except (OSError, subprocess.SubprocessError):
-        return ""
-    output = bytes(result.stdout or b"") + bytes(result.stderr or b"")
-    if result.returncode != 0 or len(output) > MAX_BACKEND_VERSION_BYTES:
-        return ""
-    match = _VERSION.search(output.decode("utf-8", errors="replace"))
-    if match is None:
-        return ""
-    try:
-        capability = click_process.run_argv(
-            [
-                executable,
-                "-D",
-                "-f",
-                "-qq",
-                "-e",
-                "trace=none",
-                "-o",
-                os.devnull,
-                "--",
-                executable,
-                "--version",
-            ],
-            stdout=subprocess.DEVNULL,
-            stderr=subprocess.DEVNULL,
-            timeout=3.0,
-        )
-    except (OSError, subprocess.SubprocessError):
-        return ""
-    return match.group(1) if capability.returncode == 0 else ""
-
-
-def _already_traced() -> bool:
-    try:
-        for line in Path("/proc/self/status").read_text(
-            encoding="utf-8", errors="strict"
-        ).splitlines():
-            if line.startswith("TracerPid:"):
-                return int(line.split(":", 1)[1].strip()) != 0
-    except (OSError, UnicodeError, ValueError):
-        return False
-    return False
-
-
-def _read_pipe(
-    read_fd: int,
-    limit: int,
-    capture: _PipeCapture,
-    stop: threading.Event,
-) -> None:
-    retained = bytearray()
-    try:
-        while True:
-            if stop.is_set():
-                capture.failed = True
-                break
-            ready, _, _ = select.select([read_fd], [], [], 0.05)
-            if not ready:
-                if stop.is_set():
-                    capture.failed = True
-                    break
-                continue
-            chunk = os.read(read_fd, 64 * 1024)
-            if not chunk:
-                break
-            remaining = max(0, limit - len(retained))
-            if remaining:
-                retained.extend(chunk[:remaining])
-            if len(chunk) > remaining:
-                capture.truncated = True
-    except OSError:
-        capture.failed = True
-    finally:
-        try:
-            os.close(read_fd)
-        except OSError:
-            pass
-    capture.data = bytes(retained)
-
-
-def _read_fifo(
-    path: Path,
-    limit: int,
-    capture: _PipeCapture,
-    stop: threading.Event,
-) -> None:
-    try:
-        read_fd = os.open(path, os.O_RDONLY)
-    except OSError:
-        capture.failed = True
-        return
-    _read_pipe(read_fd, limit, capture, stop)
-
-
-def _wake_fifo(path: Path) -> None:
-    try:
-        descriptor = os.open(path, os.O_RDWR | os.O_NONBLOCK)
-    except OSError:
-        return
-    try:
-        os.close(descriptor)
-    except OSError:
-        pass
-
-
-def _create_trace_fifo(workspace: Path) -> tuple[Path, Path] | None:
-    try:
-        workspace_root = workspace.resolve(strict=True)
-    except (OSError, RuntimeError):
-        return None
-    for raw_root in ("/tmp", "/var/tmp"):
-        directory: Path | None = None
-        fifo: Path | None = None
-        try:
-            temporary_root = Path(raw_root).resolve(strict=True)
-            if not temporary_root.is_dir():
-                continue
-            try:
-                temporary_root.relative_to(workspace_root)
-            except ValueError:
-                pass
-            else:
-                continue
-            directory = Path(
-                tempfile.mkdtemp(prefix="click-shadow-", dir=temporary_root)
-            )
-            try:
-                directory.relative_to(workspace_root)
-            except ValueError:
-                pass
-            else:
-                directory.rmdir()
-                continue
-            directory.chmod(0o700)
-            fifo = directory / "trace.pipe"
-            os.mkfifo(fifo, mode=0o600)
-            return directory, fifo
-        except OSError:
-            _remove_trace_fifo(directory, fifo)
-            continue
-    return None
-
-
-def _remove_trace_fifo(directory: Path | None, fifo: Path | None) -> None:
-    if fifo is not None:
-        try:
-            fifo.unlink(missing_ok=True)
-        except OSError:
-            pass
-    if directory is not None:
-        try:
-            directory.rmdir()
-        except OSError:
-            pass
-
-
-def _strip_pid(line: str) -> tuple[str, str]:
-    for pattern in (_PID_BRACKET, _PID_PLAIN):
-        match = pattern.match(line)
-        if match is not None:
-            return match.group(1), line[match.end() :]
-    return "root", line
-
-
-def _decoded_path(arguments: str) -> tuple[str | None, int, bool]:
-    match = _QUOTED.search(arguments)
-    if match is None:
-        return None, -1, False
-    truncated = arguments[match.end() :].lstrip().startswith("...")
-    try:
-        value = ast.literal_eval(match.group(0))
-    except (SyntaxError, ValueError):
-        return None, -1, truncated
-    if not isinstance(value, str) or "\x00" in value:
-        return None, -1, truncated
-    return value, match.start(), truncated
-
-
-def _fd_path(arguments: str, *, before: int | None = None) -> str:
-    prefix = arguments if before is None else arguments[:before]
-    matches = list(_FD_PATH.finditer(prefix))
-    return matches[-1].group(1) if matches else ""
-
-
-def _absolute_observed_path(
-    raw_path: str,
-    *,
-    arguments: str,
-    path_offset: int,
-    cwd: Path,
-) -> Path | None:
-    if not raw_path:
-        return None
-    candidate = Path(raw_path)
-    if not candidate.is_absolute():
-        annotated = _fd_path(arguments, before=path_offset)
-        base = Path(annotated) if annotated.startswith("/") else cwd
-        candidate = base / candidate
-    try:
-        return Path(os.path.normpath(os.path.abspath(candidate)))
-    except (OSError, ValueError):
-        return None
-
-
-def _return_integer(result: str) -> int | None:
-    match = _RETURN_INTEGER.match(result)
-    if match is None:
-        return None
-    try:
-        return int(match.group(1))
-    except ValueError:
-        return None
-
-
-def parse_strace(
-    raw: bytes,
-    *,
-    workspace: Path,
-    initial_cwd: Path | None = None,
-    truncated: bool = False,
-) -> ParsedTrace:
-    """Parse bounded strace text into repository-relative aggregate inputs."""
-    try:
-        root = workspace.resolve(strict=True)
-    except (OSError, RuntimeError):
-        root = Path(os.path.abspath(workspace))
-    initial = initial_cwd or root
-    try:
-        initial = Path(os.path.normpath(os.path.abspath(initial)))
-    except (OSError, ValueError):
-        initial = root
-
-    cwd_by_pid: dict[str, Path] = {"root": initial}
-    inputs: dict[str, dict[str, Any]] = {}
-    conflicts: set[str] = set()
-    external_paths: set[str] = set()
-    unresolved = 1 if truncated else 0
-    child_processes = 0
-    root_exec_observed = False
-
-    def add_path(path: Path | None, *, kind: str, operation: str) -> None:
-        nonlocal unresolved
-        if path is None:
-            unresolved = _bounded_add(unresolved, 1)
-            return
-        try:
-            relative_path = path.relative_to(root)
-        except ValueError:
-            external = os.path.normcase(str(path))
-            if external not in external_paths:
-                if (
-                    len(external_paths)
-                    >= click_dependency_cache.MAX_SHADOW_OBSERVER_INPUTS
-                ):
-                    unresolved = _bounded_add(unresolved, 1)
-                else:
-                    external_paths.add(external)
-            return
-        relative = relative_path.as_posix()
-        if relative in {"", "."}:
-            unresolved = _bounded_add(unresolved, 1)
-            return
-        if kind == "directory":
-            relative = f"{relative.rstrip('/')}/"
-        try:
-            encoded_relative = relative.encode("utf-8")
-        except UnicodeEncodeError:
-            unresolved = _bounded_add(unresolved, 1)
-            return
-        if (
-            len(encoded_relative) > click_dependency_cache.MAX_SHADOW_OBSERVER_PATH_BYTES
-            or "\\" in relative
-            or any(
-                ord(character) < 32 or ord(character) == 127
-                for character in relative
-            )
-        ):
-            unresolved = _bounded_add(unresolved, 1)
-            return
-        existing = inputs.get(relative)
-        if existing is not None and existing["kind"] != kind:
-            inputs.pop(relative, None)
-            conflicts.add(relative)
-            unresolved = _bounded_add(unresolved, 1)
-            return
-        if relative in conflicts:
-            return
-        if existing is None:
-            if len(inputs) >= click_dependency_cache.MAX_SHADOW_OBSERVER_INPUTS:
-                unresolved = _bounded_add(unresolved, 1)
-                return
-            inputs[relative] = {
-                "path": relative,
-                "kind": kind,
-                "operations": {operation},
-            }
-        else:
-            existing["operations"].add(operation)
-
-    for raw_line in raw.decode("utf-8", errors="replace").splitlines():
-        pid, line = _strip_pid(raw_line.strip())
-        if not line or line.startswith(("--- ", "+++ ", "strace:")):
-            continue
-        if "<unfinished ...>" in line or line.startswith("<..."):
-            unresolved = _bounded_add(unresolved, 1)
-            continue
-        match = _CALL.match(line)
-        if match is None:
-            unresolved = _bounded_add(unresolved, 1)
-            continue
-        call, arguments, result = match.groups()
-        cwd = cwd_by_pid.get(pid, initial)
-        returned = _return_integer(result)
-
-        if call in _CHILD_CALLS:
-            if returned is not None and returned > 0:
-                child_processes = _bounded_add(child_processes, 1)
-                if (
-                    len(cwd_by_pid)
-                    <= click_dependency_cache.MAX_SHADOW_OBSERVER_INPUTS
-                ):
-                    cwd_by_pid[str(returned)] = cwd
-                else:
-                    unresolved = _bounded_add(unresolved, 1)
-            continue
-        if call in _IGNORED_PROCESS_CALLS:
-            continue
-        if call == "chdir":
-            path_text, offset, shortened = _decoded_path(arguments)
-            if shortened or path_text is None or returned != 0:
-                if shortened or path_text is None:
-                    unresolved = _bounded_add(unresolved, 1)
-                continue
-            changed = _absolute_observed_path(
-                path_text,
-                arguments=arguments,
-                path_offset=offset,
-                cwd=cwd,
-            )
-            if changed is None:
-                unresolved = _bounded_add(unresolved, 1)
-            else:
-                cwd_by_pid[pid] = changed
-            continue
-        if call == "fchdir":
-            annotated = _fd_path(arguments)
-            if returned == 0 and annotated.startswith("/"):
-                cwd_by_pid[pid] = Path(os.path.normpath(annotated))
-            elif returned == 0:
-                unresolved = _bounded_add(unresolved, 1)
-            continue
-        if call in _DIRECTORY_CALLS:
-            annotated = _fd_path(arguments)
-            if returned is not None and returned >= 0 and annotated.startswith("/"):
-                add_path(
-                    Path(os.path.normpath(annotated)),
-                    kind="directory",
-                    operation="enumerate",
-                )
-            elif returned is None:
-                unresolved = _bounded_add(unresolved, 1)
-            continue
-        if call not in _EXEC_CALLS | _OPEN_CALLS | _METADATA_CALLS:
-            unresolved = _bounded_add(unresolved, 1)
-            continue
-
-        path_text, offset, shortened = _decoded_path(arguments)
-        if shortened or path_text is None:
-            unresolved = _bounded_add(unresolved, 1)
-            continue
-        observed = _absolute_observed_path(
-            path_text,
-            arguments=arguments,
-            path_offset=offset,
-            cwd=cwd,
-        )
-        missing = result.startswith("-1 ENOENT")
-        if call in _EXEC_CALLS:
-            if returned == 0:
-                root_exec_observed = True
-                add_path(observed, kind="file", operation="execute")
-            elif not missing:
-                unresolved = _bounded_add(unresolved, 1)
-            else:
-                add_path(observed, kind="missing", operation="execute")
-            continue
-        if missing:
-            operation = "read" if call in _OPEN_CALLS else "metadata"
-            add_path(observed, kind="missing", operation=operation)
-            continue
-        if returned is None or returned < 0:
-            continue
-        if call in _OPEN_CALLS:
-            if "O_WRONLY" in arguments and "O_RDWR" not in arguments:
-                continue
-            resolved_descriptor = _fd_path(result)
-            if resolved_descriptor.startswith("/"):
-                observed = Path(os.path.normpath(resolved_descriptor))
-            directory = "O_DIRECTORY" in arguments
-            add_path(
-                observed,
-                kind="directory" if directory else "file",
-                operation="metadata" if "O_PATH" in arguments or directory else "read",
-            )
-            continue
-        directory = "S_IFDIR" in arguments
-        add_path(
-            observed,
-            kind="directory" if directory else "file",
-            operation="metadata",
-        )
-
-    normalized_inputs = tuple(
-        {
-            "path": relative,
-            "kind": inputs[relative]["kind"],
-            "operations": sorted(inputs[relative]["operations"]),
-        }
-        for relative in sorted(inputs)
-    )
-    process_tree_complete = not truncated and unresolved == 0
-    return ParsedTrace(
-        inputs=normalized_inputs,
-        external_input_count=min(
-            len(external_paths), click_dependency_cache.MAX_JSON_SAFE_INTEGER
-        ),
-        unresolved_event_count=unresolved,
-        child_process_count=child_processes,
-        process_tree_complete=process_tree_complete,
-        root_exec_observed=root_exec_observed,
-    )
-
-
-def _record(
-    *,
-    evidence_key: str,
-    check_digest: str,
-    mutation_revision: int,
-    backend_name: str | None,
-    backend_version: str = "",
-    backend_digest: str = "",
-    inputs: Sequence[dict[str, Any]] = (),
-    status: str,
-    external_input_count: int = 0,
-    unresolved_event_count: int = 0,
-    child_process_count: int = 0,
-    process_tree_complete: bool,
-    command_duration_ms: int,
-    observer_overhead_ms: int,
-) -> dict[str, Any]:
-    safe_duration = max(
-        0,
-        min(command_duration_ms, click_dependency_cache.MAX_JSON_SAFE_INTEGER),
-    )
-    safe_overhead = max(0, min(observer_overhead_ms, safe_duration))
-    try:
-        return click_dependency_cache.shadow_observer_record(
-            evidence_key=evidence_key,
-            check_digest=check_digest,
-            mutation_revision=mutation_revision,
-            backend_name=backend_name,
-            backend_version=backend_version,
-            backend_digest=backend_digest,
-            inputs=inputs,
-            status=status,
-            external_input_count=external_input_count,
-            unresolved_event_count=unresolved_event_count,
-            child_process_count=child_process_count,
-            process_tree_complete=process_tree_complete,
-            command_duration_ms=safe_duration,
-            observer_overhead_ms=safe_overhead,
-        )
-    except (TypeError, ValueError):
-        return click_dependency_cache.shadow_observer_record(
-            evidence_key=evidence_key,
-            check_digest=check_digest,
-            mutation_revision=mutation_revision,
-            backend_name=None,
-            status="unavailable",
-            process_tree_complete=False,
-            command_duration_ms=safe_duration,
-            observer_overhead_ms=safe_overhead,
-        )
-
-
-def _fallback_execution(
-    execute_unobserved: FallbackExecutor,
-    *,
-    evidence_key: str,
-    check_digest: str,
-    mutation_revision: int,
-    status: str,
-    backend_name: str | None = None,
-    backend_version: str = "",
-    backend_digest: str = "",
-    preparation_ms: int = 0,
-) -> ShadowExecution:
-    command_started = time.monotonic()
-    exit_code = int(execute_unobserved())
-    duration_ms = max(0, int((time.monotonic() - command_started) * 1000))
-    return ShadowExecution(
-        exit_code=exit_code,
-        record=_record(
-            evidence_key=evidence_key,
-            check_digest=check_digest,
-            mutation_revision=mutation_revision,
-            backend_name=backend_name,
-            backend_version=backend_version,
-            backend_digest=backend_digest,
-            status=status,
-            process_tree_complete=False,
-            command_duration_ms=duration_ms,
-            observer_overhead_ms=preparation_ms,
-        ),
-    )
-
-
-def run_unobserved(
-    execute_unobserved: FallbackExecutor,
-    *,
-    evidence_key: str,
-    check_digest: str,
-    mutation_revision: int,
-) -> ShadowExecution:
-    """Execute through the established path and record unavailable telemetry."""
-    return _fallback_execution(
-        execute_unobserved,
-        evidence_key=evidence_key,
-        check_digest=check_digest,
-        mutation_revision=mutation_revision,
-        status="unavailable",
-    )
+# Keep the intentionally tested Linux compatibility helpers on the old facade.
+probe_strace_version = click_observer_linux.probe_strace_version
+parse_strace = click_observer_linux.parse_strace
+_already_traced = click_observer_linux._already_traced
+_file_digest = click_observer_linux._file_digest
+click_process = click_observer_linux.click_process
 
 
 def run_command(
@@ -693,405 +76,48 @@ def run_command(
     resolve_backend: BackendResolver,
     digest_file: FileDigester = _file_digest,
     probe_version: BackendProbe = probe_strace_version,
-    spawn_argv: SpawnArgv = click_process.spawn_argv,
+    spawn_argv: SpawnArgv = click_observer_linux.click_process.spawn_argv,
     terminate_group: Callable[[subprocess.Popen[Any]], int] = (
-        click_process.terminate_process_group
+        click_observer_linux.click_process.terminate_process_group
     ),
     system_name: str | None = None,
     already_traced: Callable[[], bool] = _already_traced,
     capture_limit: int = MAX_RAW_TRACE_BYTES,
 ) -> ShadowExecution:
-    """Execute one command exactly once and collect best-effort Linux trace data."""
+    """Select one backend and execute the target exactly once."""
+
     try:
         system = platform.system() if system_name is None else system_name
-        nested_trace = already_traced() if system == "Linux" else False
+        capability = select_backend(system)
     except Exception:
-        system = ""
-        nested_trace = False
-    if system != "Linux" or nested_trace:
-        return _fallback_execution(
-            execute_unobserved,
-            evidence_key=evidence_key,
-            check_digest=check_digest,
-            mutation_revision=mutation_revision,
-            status="unavailable",
-        )
-
-    preparation_started = time.monotonic()
-    try:
-        executable, resolve_error = resolve_backend("strace", workspace=workspace)
-    except Exception:
-        executable, resolve_error = None, "backend resolution failed"
-    if resolve_error or not isinstance(executable, str) or not executable:
-        preparation_ms = max(0, int((time.monotonic() - preparation_started) * 1000))
-        return _fallback_execution(
-            execute_unobserved,
-            evidence_key=evidence_key,
-            check_digest=check_digest,
-            mutation_revision=mutation_revision,
-            status="unavailable",
-            preparation_ms=preparation_ms,
-        )
-    try:
-        digest = digest_file(Path(executable))
-        version = probe_version(executable)
-    except Exception:
-        digest = ""
-        version = ""
-    if (
-        not isinstance(digest, str)
-        or _DIGEST.fullmatch(digest) is None
-        or not isinstance(version, str)
-        or re.fullmatch(r"[A-Za-z0-9][A-Za-z0-9._+-]{0,63}", version) is None
-    ):
-        preparation_ms = max(0, int((time.monotonic() - preparation_started) * 1000))
-        return _fallback_execution(
-            execute_unobserved,
-            evidence_key=evidence_key,
-            check_digest=check_digest,
-            mutation_revision=mutation_revision,
-            status="unavailable",
-            preparation_ms=preparation_ms,
-        )
-    preparation_ms = max(0, int((time.monotonic() - preparation_started) * 1000))
-
-    capture = _PipeCapture()
-    trace_directory: Path | None = None
-    trace_fifo: Path | None = None
-    child: subprocess.Popen[Any] | None = None
-    reader: threading.Thread | None = None
-    stop_reader = threading.Event()
-    command_started = time.monotonic()
-    try:
-        location = _create_trace_fifo(workspace)
-        if location is None:
-            raise OSError("private external trace FIFO is unavailable")
-        trace_directory, trace_fifo = location
-        bounded_capture_limit = (
-            capture_limit
-            if isinstance(capture_limit, int)
-            and not isinstance(capture_limit, bool)
-            and capture_limit > 0
-            else MAX_RAW_TRACE_BYTES
-        )
-        reader = threading.Thread(
-            target=_read_fifo,
-            args=(trace_fifo, bounded_capture_limit, capture, stop_reader),
-            daemon=True,
-        )
-        reader.start()
-        traced_argv = [
-            executable,
-            "-D",
-            "-f",
-            "-qq",
-            "-yy",
-            "-s",
-            str(STRACE_STRING_LIMIT),
-            "-e",
-            STRACE_TRACE_EXPRESSION,
-            "-o",
-            str(trace_fifo),
-            "--",
-            *list(argv),
-        ]
-        child = spawn_argv(
-            traced_argv,
-            cwd=workspace,
-            env=dict(environment),
-        )
-        exit_code = int(child.wait())
-    except KeyboardInterrupt:
-        if child is not None:
-            try:
-                terminate_group(child)
-            except Exception:
-                pass
-        exit_code = 130
-        capture.failed = True
-    except (OSError, subprocess.SubprocessError, RuntimeError, TypeError, ValueError):
-        if child is None:
-            if reader is not None:
-                stop_reader.set()
-                if trace_fifo is not None:
-                    _wake_fifo(trace_fifo)
-                reader.join(timeout=1)
-            preparation_ms = _bounded_add(
-                preparation_ms,
-                max(0, int((time.monotonic() - command_started) * 1000)),
-            )
-            return _fallback_execution(
-                execute_unobserved,
-                evidence_key=evidence_key,
-                check_digest=check_digest,
-                mutation_revision=mutation_revision,
-                status="failed",
-                backend_name="strace",
-                backend_version=version,
-                backend_digest=digest,
-                preparation_ms=preparation_ms,
-            )
-        try:
-            exit_code = int(child.wait())
-        except (
-            OSError,
-            subprocess.SubprocessError,
-            RuntimeError,
-            TypeError,
-            ValueError,
-        ):
-            exit_code = 127
-        capture.failed = True
-    finally:
-        if reader is not None:
-            reader.join(timeout=1)
-            if reader.is_alive():
-                capture.failed = True
-                stop_reader.set()
-                if trace_fifo is not None:
-                    _wake_fifo(trace_fifo)
-                reader.join(timeout=0.2)
-        _remove_trace_fifo(trace_directory, trace_fifo)
-
-    command_duration_ms = max(0, int((time.monotonic() - command_started) * 1000))
-    parsing_started = time.monotonic()
-    raw_trace = capture.data
-    capture.data = b""
-    try:
-        parsed = parse_strace(
-            raw_trace,
-            workspace=observation_root or workspace,
-            initial_cwd=workspace,
-            truncated=capture.truncated or capture.failed,
-        )
-    except Exception:
-        capture.failed = True
-        parsed = ParsedTrace(
-            inputs=(),
-            external_input_count=0,
-            unresolved_event_count=1,
-            child_process_count=0,
-            process_tree_complete=False,
-            root_exec_observed=False,
-        )
-    finally:
-        raw_trace = b""
-    parsing_ms = max(0, int((time.monotonic() - parsing_started) * 1000))
-    observer_overhead_ms = _bounded_add(preparation_ms, parsing_ms)
-
-    identity_started = time.monotonic()
-    try:
-        final_digest = digest_file(Path(executable))
-    except Exception:
-        final_digest = ""
-    identity_ms = max(0, int((time.monotonic() - identity_started) * 1000))
-    observer_overhead_ms = _bounded_add(observer_overhead_ms, identity_ms)
-    if final_digest != digest:
-        return ShadowExecution(
-            exit_code=exit_code,
-            record=_record(
-                evidence_key=evidence_key,
-                check_digest=check_digest,
-                mutation_revision=mutation_revision,
-                backend_name=None,
-                status="unavailable",
-                process_tree_complete=False,
-                command_duration_ms=command_duration_ms,
-                observer_overhead_ms=observer_overhead_ms,
-            ),
-        )
-
-    status = (
-        "failed"
-        if capture.failed or not parsed.root_exec_observed
-        else "partial"
-        if capture.truncated or not parsed.process_tree_complete
-        else "complete"
-    )
-    record = _record(
-        evidence_key=evidence_key,
-        check_digest=check_digest,
-        mutation_revision=mutation_revision,
-        backend_name="strace",
-        backend_version=version,
-        backend_digest=digest,
-        inputs=parsed.inputs,
-        status=status,
-        external_input_count=parsed.external_input_count,
-        unresolved_event_count=parsed.unresolved_event_count,
-        child_process_count=parsed.child_process_count,
-        process_tree_complete=(
-            parsed.process_tree_complete and status == "complete"
-        ),
-        command_duration_ms=command_duration_ms,
-        observer_overhead_ms=observer_overhead_ms,
-    )
-    return ShadowExecution(exit_code=exit_code, record=record)
-
-
-def combine_records(
-    records: Sequence[dict[str, Any]],
-    *,
-    evidence_key: str,
-    check_digest: str,
-    mutation_revision: int,
-    unexecuted_checks: int = 0,
-) -> dict[str, Any] | None:
-    """Combine per-command records into the latest bounded source record."""
-    valid = [
-        record
-        for record in records
-        if click_dependency_cache.shadow_observer_record_is_valid(record)
-        and record["binding"]
-        == {
-            "evidence_key": evidence_key,
-            "check_digest": check_digest,
-            "mutation_revision": mutation_revision,
-        }
-    ]
-    if len(valid) != len(records) or not valid:
-        return None
-    backends = {
-        tuple(sorted(record["backend"].items()))
-        for record in valid
-        if isinstance(record.get("backend"), dict)
-    }
-    if len(backends) > 1:
-        return None
-    backend = dict(next(iter(backends))) if backends else None
-    observed = [record for record in valid if record["status"] != "unavailable"]
-    duration = 0
-    overhead = 0
-    for record in valid:
-        duration = _bounded_add(duration, int(record["command_duration_ms"]))
-        overhead = _bounded_add(overhead, int(record["observer_overhead_ms"]))
-    if not observed:
-        return _record(
-            evidence_key=evidence_key,
-            check_digest=check_digest,
-            mutation_revision=mutation_revision,
+        capability = BackendCapability(
+            system="",
             backend_name=None,
             status="unavailable",
-            process_tree_complete=False,
-            command_duration_ms=duration,
-            observer_overhead_ms=overhead,
+            reason="capability-detection-failed",
         )
-    assert backend is not None
-    unavailable_count = len(valid) - len(observed)
-    status = (
-        "failed"
-        if any(record["status"] == "failed" for record in observed)
-        else "partial"
-        if unavailable_count
-        or unexecuted_checks
-        or any(record["status"] == "partial" for record in observed)
-        else "complete"
-    )
-    unresolved = unexecuted_checks + unavailable_count
-    external = 0
-    children = 0
-    inputs: list[dict[str, Any]] = []
-    for record in observed:
-        unresolved = _bounded_add(
-            unresolved, int(record["unresolved_event_count"])
+    if capability.status != "available" or capability.backend_name != "strace":
+        return run_unobserved(
+            execute_unobserved,
+            evidence_key=evidence_key,
+            check_digest=check_digest,
+            mutation_revision=mutation_revision,
         )
-        external = _bounded_add(
-            external, int(record["external_input_count"])
-        )
-        children = _bounded_add(children, int(record["child_process_count"]))
-        inputs.extend(record["inputs"])
-    process_tree_complete = bool(
-        status == "complete"
-        and all(record["process_tree_complete"] for record in observed)
-    )
-    return _record(
+    return click_observer_linux.run_command(
+        argv,
+        workspace=workspace,
+        observation_root=observation_root,
+        environment=environment,
         evidence_key=evidence_key,
         check_digest=check_digest,
         mutation_revision=mutation_revision,
-        backend_name=str(backend["name"]),
-        backend_version=str(backend["version"]),
-        backend_digest=str(backend["digest"]),
-        inputs=inputs,
-        status=status,
-        external_input_count=external,
-        unresolved_event_count=unresolved,
-        child_process_count=children,
-        process_tree_complete=process_tree_complete,
-        command_duration_ms=duration,
-        observer_overhead_ms=overhead,
-    )
-
-
-def fresh_state() -> dict[str, Any]:
-    return {"version": SHADOW_STATE_VERSION, "records": {}}
-
-
-def state_is_valid(value: Any) -> bool:
-    if not isinstance(value, dict) or set(value) != {"version", "records"}:
-        return False
-    if value.get("version") != SHADOW_STATE_VERSION:
-        return False
-    records = value.get("records")
-    if (
-        not isinstance(records, dict)
-        or len(records) > MAX_SHADOW_STATE_RECORDS
-    ):
-        return False
-    return all(
-        isinstance(key, str)
-        and _DIGEST.fullmatch(key) is not None
-        and click_dependency_cache.shadow_observer_record_is_valid(record)
-        and record["binding"]["evidence_key"] == key
-        for key, record in records.items()
-    )
-
-
-def records_from_verification(verification: Any) -> dict[str, dict[str, Any]]:
-    if not isinstance(verification, dict):
-        return {}
-    state = verification.get(SHADOW_STATE_FIELD)
-    if not state_is_valid(state):
-        return {}
-    return {
-        key: dict(record)
-        for key, record in state["records"].items()
-    }
-
-
-def store_records(
-    verification: dict[str, Any], records: Mapping[str, Any]
-) -> int:
-    """Persist only canonical latest records; malformed telemetry is ignored."""
-    current = records_from_verification(verification)
-    stored = 0
-    for key, record in sorted(records.items()):
-        if (
-            len(current) >= MAX_SHADOW_STATE_RECORDS
-            and key not in current
-        ):
-            continue
-        if (
-            not isinstance(key, str)
-            or _DIGEST.fullmatch(key) is None
-            or not click_dependency_cache.shadow_observer_record_is_valid(record)
-            or record["binding"]["evidence_key"] != key
-        ):
-            continue
-        current[key] = record
-        stored += 1
-    verification[SHADOW_STATE_FIELD] = {
-        "version": SHADOW_STATE_VERSION,
-        "records": current,
-    }
-    return stored
-
-
-def advisory(record: Any) -> str:
-    if not click_dependency_cache.shadow_observer_record_is_valid(record):
-        return "[Click shadow] failed: canonical record unavailable"
-    return (
-        f"[Click shadow] {record['status']}: inputs={len(record['inputs'])} "
-        f"external={record['external_input_count']} "
-        f"unresolved={record['unresolved_event_count']} "
-        f"overhead={record['observer_overhead_ms']}ms"
+        execute_unobserved=execute_unobserved,
+        resolve_backend=resolve_backend,
+        digest_file=digest_file,
+        probe_version=probe_version,
+        spawn_argv=spawn_argv,
+        terminate_group=terminate_group,
+        system_name=system,
+        already_traced=already_traced,
+        capture_limit=capture_limit,
     )
