+++ scripts/build_antigravity_distribution.py
@@ -24,6 +24,9 @@ HOOK_FILES = (
     "click_change_policy.py",
     "click_dependency_cache.py",
     "click_dependency_trace.py",
+    "click_observer_backend.py",
+    "click_observer_common.py",
+    "click_observer_linux.py",
     "click_evidence.py",
     "click_evidence_shards.py",
     "click_receipt.py",
