+++ tests/repository_policy_core.py
@@ -460,6 +460,9 @@ class RepositoryPolicyTests(unittest.TestCase):
         self.assertIn("hooks/click_browser_advisory.py", classification)
         self.assertIn("hooks/click_dependency_cache.py", classification)
         self.assertIn("hooks/click_dependency_trace.py", classification)
+        self.assertIn("hooks/click_observer_backend.py", classification)
+        self.assertIn("hooks/click_observer_common.py", classification)
+        self.assertIn("hooks/click_observer_linux.py", classification)
         self.assertIn("hooks/click_evidence_shards.py", classification)
         self.assertIn("hooks/click_shadow_intelligence.py", classification)
         self.assertIn("hooks/click_shadow_dashboard.py", classification)
@@ -551,6 +554,9 @@ class RepositoryPolicyTests(unittest.TestCase):
                 "click_contract_state.py",
                 "click_dependency_cache.py",
                 "click_dependency_trace.py",
+                "click_observer_backend.py",
+                "click_observer_common.py",
+                "click_observer_linux.py",
                 "click_evidence_shards.py",
                 "click_shadow_dashboard.py",
                 "click_shadow_intelligence.py",
